# Documentation/src/content/schemas/swagger.yaml
openapi: 3.0.0
info:
  title: Zello API
  version: 1.0.0
  description: |-
    A comprehensive API for managing projects, lists, and cards similar to Trello. This API allows users to:
    - Create and manage workspaces, projects, lists, and cards
    - Organize work using customizable boards and lists
    - Collaborate with team members through comments and mentions
    - Track work progress with activity logs
    - Manage permissions and roles at different levels
    
    ### Features
    - Full CRUD operations for all resources
    - Rich card features including attachments, checklists, and labels
    - Comprehensive activity tracking
    - Fine-grained permission system
    - Authentication and user management
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workspace:
      description: Represents an organization or team workspace that contains multiple projects
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174010
        name:
          type: string
          example: Engineering Team
        description:
          type: string
          example: Main workspace for the engineering department
        ownerId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        settings:
          type: object
          properties:
            defaultTaskTemplate:
              type: object
            defaultIssueTemplate:
              type: object
            allowedLabels:
              type: array
              items:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Project:
      description: Represents a project with its lists and settings
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
          example: "User Authentication Service"
        description:
          type: string
        type:
          type: string
          enum:
            - scrum
            - kanban
            - basic
        startDate:
          type: string
          format: date-time
          description: Project start date
          example: "2024-01-01T00:00:00Z"
        endDate:
          type: string
          format: date-time
          description: Project target completion date
          example: "2024-06-30T23:59:59Z"
        settings:
          type: object
          properties:
            sprintDuration:
              type: integer
            pointScale:
              type: string
            defaultEstimateUnit:
              type: string
            defaultIssueTypes:
              type: array
              description: Default issue types available in this project
              items:
                type: string
                enum: [ task, bug, feature, improvement, epic ]
        status:
          type: string
          enum:
            - planning    # Project is in planning phase
            - active     # Project is currently being worked on
            - onHold    # Project is temporarily paused
            - completed # Project has been completed
            - archived  # Project is archived
        actualStartDate:
          type: string
          format: date-time
          description: Actual date when project started
        actualEndDate:
          type: string
          format: date-time
          description: Actual date when project was completed
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Invalid request parameters
        code:
          type: string
          example: INVALID_PARAMETERS
security:
  - bearerAuth: [ ]
paths:
...
  /workspaces/{workspaceId}/custom-roles:
    post:
      summary: Create a custom role
      description: Creates a new custom role within a workspace with specified permissions
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the workspace to create the custom role in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - permissions
              properties:
                name:
                  type: string
                  example: "Technical Lead"
                  minLength: 1
                  maxLength: 50
                description:
                  type: string
                  example: "Role for technical team leaders"
                permissions:
                  type: array
                  items:
                    type: string
                    enum:
                      - manage_workspace
                      - manage_members
                      - create_project
                      - manage_lists
                      - manage_cards
                      - comment
                      - view
            example:
              name: "Technical Lead"
              description: "Role for technical team leaders"
              permissions: [
                "create_project",
                "manage_lists",
                "manage_cards",
                "comment",
                "view"
              ]
      responses:
        '201':
          description: Custom role successfully created
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidName:
                  value:
                    message: "Role name must be between 1 and 50 characters"
                    code: "INVALID_NAME"
                invalidPermissions:
                  value:
                    message: "Invalid permissions specified"
                    code: "INVALID_PERMISSIONS"
                duplicateName:
                  value:
                    message: "Role name already exists in workspace"
                    code: "DUPLICATE_NAME"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions to create custom roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Only workspace admins can create custom roles"
                code: "INSUFFICIENT_PERMISSIONS"
        '404':
          description: Workspace not found
        '409':
          description: Conflict - maximum number of custom roles reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Maximum number of custom roles (10) reached for workspace"
                code: "MAX_ROLES_REACHED"
...
  /workspaces:
    post:
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                settings:
                  type: object
                  properties:
                    defaultTaskTemplate:
                      type: object
                    defaultIssueTemplate:
                      type: object
                    allowedLabels:
                      type: array
                      items:
                        type: string
            example:
              name: "Product Development"
              description: "Main workspace for product development team"
              settings: {
                defaultTaskTemplate: {
                  assignee: null,
                  labels: [ "feature" ],
                  checklist: [ "Planning", "Development", "Testing" ]
                },
                defaultIssueTemplate: {
                  priority: "medium",
                  labels: [ "bug" ]
                },
                allowedLabels: [ "bug", "feature", "improvement", "urgent" ]
              }

      responses:
        '201':
          description: Workspace created successfully
        '400':
          description: Invalid request parameters
        '401':
          description: Unauthorized
  /workspaces/{workspaceId}/projects:
    get:
      summary: List projects in workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [ active, archived, completed ]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
              example:
                items: [
                  {
                    id: "123e4567-e89b-12d3-a456-426614174030",
                    workspaceId: "123e4567-e89b-12d3-a456-426614174010",
                    name: "Website Redesign",
                    description: "Complete overhaul of company website with modern design and improved UX",
                    type: "kanban",
                    settings: {
                      sprintDuration: 14,
                      pointScale: "fibonacci",
                      defaultEstimateUnit: "days"
                    },
                    status: "active",
                    createdAt: "2024-01-01T12:00:00Z",
                    updatedAt: "2024-01-01T12:00:00Z"
                  },
                  {
                    id: "123e4567-e89b-12d3-a456-426614174031",
                    workspaceId: "123e4567-e89b-12d3-a456-426614174010",
                    name: "Mobile App Development",
                    description: "Native mobile application for iOS and Android platforms",
                    type: "scrum",
                    settings: {
                      sprintDuration: 14,
                      pointScale: "fibonacci",
                      defaultEstimateUnit: "hours"
                    },
                    status: "active",
                    createdAt: "2024-01-02T12:00:00Z",
                    updatedAt: "2024-01-02T12:00:00Z"
                  },
                  {
                    id: "123e4567-e89b-12d3-a456-426614174032",
                    workspaceId: "123e4567-e89b-12d3-a456-426614174010",
                    name: "API Integration",
                    description: "Third-party API integration project",
                    type: "basic",
                    settings: {
                      defaultEstimateUnit: "hours"
                    },
                    status: "completed",
                    createdAt: "2023-12-01T12:00:00Z",
                    updatedAt: "2024-01-15T16:30:00Z"
                  }
                ]
                total: 3,
                page: 1,
                pages: 1
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Invalid page number. Page must be greater than 0"
                code: "INVALID_PAGE_NUMBER"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Authentication required"
                code: "UNAUTHORIZED"
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "You don't have permission to view projects in this workspace"
                code: "INSUFFICIENT_PERMISSIONS"
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Workspace not found"
                code: "WORKSPACE_NOT_FOUND"
  /workspaces/{workspaceId}:
    get:
      summary: Get workspace details
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
              example:
                id: "123e4567-e89b-12d3-a456-426614174010"
                name: "Engineering Team"
                description: "Main workspace for engineering department"
                ownerId: "123e4567-e89b-12d3-a456-426614174000"
                settings: {
                  defaultTaskTemplate: {
                    assignee: null,
                    labels: [ "feature" ],
                    checklist: [ "Planning", "Development", "Testing" ]
                  },
                  defaultIssueTemplate: {
                    priority: "medium",
                    labels: [ "bug" ],
                    assignee: null
                  },
                  allowedLabels: [ "bug", "feature", "improvement", "urgent" ]
                }
                createdAt: "2024-01-01T12:00:00Z"
                updatedAt: "2024-01-15T08:30:00Z"
        '404':
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Workspace not found"
                code: "WORKSPACE_NOT_FOUND"

    put:
      summary: Update workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Product Development"
              description: "Main workspace for product development team"
              settings: {
                defaultTaskTemplate: {
                  assignee: null,
                  labels: [ "feature" ],
                  checklist: [ "Planning", "Development", "Testing" ]
                },
                allowedLabels: [ "bug", "feature", "improvement", "urgent" ]
              }
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                settings:
                  type: object
                  properties:
                    defaultTaskTemplate:
                      type: object
                    defaultIssueTemplate:
                      type: object
                    allowedLabels:
                      type: array
                      items:
                        type: string
      responses:
        '200':
          description: Workspace updated
        '403':
          description: Insufficient permissions
        '404':
          description: Workspace not found

    delete:
      summary: Delete workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Workspace deleted
        '403':
          description: Insufficient permissions
        '404':
          description: Workspace not found
  /workspaces/{workspaceId}/members:
    post:
      summary: Add or update workspace member
      description: Add a new member or update existing member's role (cannot elevate above own role level)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - role
              properties:
                userId:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum:
                    - admin
                    - user
                    - guest
            example:
              userId: "123e4567-e89b-12d3-a456-426614174000"
              role: "admin"
              permissions: [
                "manage_workspace",
                "manage_members",
                "create_project",
                "manage_lists",
                "manage_cards"
              ]

      responses:
        '201':
          description: Member successfully added or updated
        '403':
          description: Insufficient permissions or attempting to elevate above own role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /workspaces/{workspaceId}/labels:
    get:
      summary: Get workspace labels
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of workspace labels
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        color:
                          type: string
                        workspaceId:
                          type: string
                          format: uuid
              examples:
                success:
                  summary: List of workspace labels
                  value: {
                    "items": [
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174050",
                        "name": "bug",
                        "color": "#FF0000",
                        "workspaceId": "123e4567-e89b-12d3-a456-426614174010"
                      },
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174051",
                        "name": "feature",
                        "color": "#00FF00",
                        "workspaceId": "123e4567-e89b-12d3-a456-426614174010"
                      },
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174052",
                        "name": "improvement",
                        "color": "#0000FF",
                        "workspaceId": "123e4567-e89b-12d3-a456-426614174010"
                      },
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174053",
                        "name": "urgent",
                        "color": "#FF4500",
                        "workspaceId": "123e4567-e89b-12d3-a456-426614174010"
                      }
                    ],
                    "total": 4,
                    "page": 1,
                    "pages": 1
                  }
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Invalid workspace ID format"
                code: "INVALID_WORKSPACE_ID"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Authentication required"
                code: "UNAUTHORIZED"
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Insufficient permissions to view workspace labels"
                code: "INSUFFICIENT_PERMISSIONS"
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Workspace not found"
                code: "WORKSPACE_NOT_FOUND"
...