# src/Zello.Api/Controllers/WorkspaceController.cs
using Microsoft.AspNetCore.Mvc;
using Newtonsoft.Json;

namespace Zello.Api.Controllers;

[ApiController]
[Route("api/v1/[controller]")]
public sealed class WorkspacesController : ControllerBase {
    [HttpPost]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult CreateWorkspace([FromBody] object workspace) {
        var response = new { Message = "Workspace created" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpGet("{workspaceId}")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult GetWorkspace(string workspaceId) {
        var response = new { Message = "Workspace retrieved" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpPut("{workspaceId}")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult UpdateWorkspace(string workspaceId, [FromBody] object workspace) {
        var response = new { Message = "Workspace updated" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpDelete("{workspaceId}")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult DeleteWorkspace(string workspaceId) {
        var response = new { Message = "Workspace deleted" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpGet("{workspaceId}/projects")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult GetWorkspaceProjects(string workspaceId) {
        var response = new { Message = "Workspace projects retrieved" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpGet("{workspaceId}/labels")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult GetWorkspaceLabels(string workspaceId) {
        var response = new { Message = "Workspace labels retrieved" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpPost("{workspaceId}/members")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult AddWorkspaceMember(string workspaceId, [FromBody] object member) {
        var response = new { Message = "Workspace member added" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }

    [HttpPost("{workspaceId}/custom-roles")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    public ContentResult CreateCustomRole(string workspaceId, [FromBody] object customRole) {
        var response = new { Message = "Custom role created" };
        return new ContentResult {
            Content = JsonConvert.SerializeObject(response),
            ContentType = "application/json",
            StatusCode = StatusCodes.Status200OK
        };
    }
}